+++ lib/src/Message.dart
@@ -8,6 +8,7 @@ import 'SIPMessage.dart';
 import 'UA.dart';
 import 'Utils.dart' as Utils;
 import 'event_manager/event_manager.dart';
+import 'event_manager/internal_events.dart';
 import 'logger.dart';
 
 class Message extends EventManager {
