+++ lib/src/RTCSession.dart
@@ -13,8 +13,8 @@ import 'RTCSession/DTMF.dart' as RTCSession_DTMF;
 import 'RTCSession/DTMF.dart';
 import 'RTCSession/Info.dart' as RTCSession_Info;
 import 'RTCSession/Info.dart';
-import 'RTCSession/ReferNotifier.dart' as RTCSession_ReferNotifier;
-import 'RTCSession/ReferSubscriber.dart' as RTCSession_ReferSubscriber;
+import 'RTCSession/ReferNotifier.dart';
+import 'RTCSession/ReferSubscriber.dart';
 import 'RequestSender.dart';
 import 'SIPMessage.dart';
 
@@ -23,6 +23,7 @@ import 'UA.dart';
 import 'URI.dart';
 import 'Utils.dart' as Utils;
 import 'event_manager/event_manager.dart';
+import 'event_manager/internal_events.dart';
 import 'logger.dart';
 import 'transactions/transaction_base.dart';
 
@@ -600,7 +601,7 @@ class RTCSession extends EventManager {
             DartSIP_C.causes.USER_DENIED_MEDIA_ACCESS,
             'User Denied Media Access');
         logger.error('emit "getusermediafailed" [error:${error.toString()}]');
-        this.emit(EventGetusermediafailed(exception: error));
+        this.emit(EventGetUserMediaFailed(exception: error));
         throw new Exceptions.InvalidStateError('getUserMedia() failed');
       }
     }
@@ -1161,7 +1162,7 @@ class RTCSession extends EventManager {
   /**
    * Refer
    */
-  refer(target, [options]) {
+  ReferSubscriber refer(target, [options]) {
     logger.debug('refer()');
 
     options = options ?? {};
@@ -1170,7 +1171,7 @@ class RTCSession extends EventManager {
 
     if (this._status != C.STATUS_WAITING_FOR_ACK &&
         this._status != C.STATUS_CONFIRMED) {
-      return false;
+      return null;
     }
 
     // Check target validity.
@@ -1179,7 +1180,7 @@ class RTCSession extends EventManager {
       throw new Exceptions.TypeError('Invalid target: ${originalTarget}');
     }
 
-    var referSubscriber = new RTCSession_ReferSubscriber.ReferSubscriber(this);
+    var referSubscriber = new ReferSubscriber(this);
 
     referSubscriber.sendRefer(target, options);
 
@@ -1189,13 +1190,14 @@ class RTCSession extends EventManager {
     this._referSubscribers[id] = referSubscriber;
 
     // Listen for ending events so we can remove it from the map.
-    referSubscriber.on(EventRequestFailed(), (EventRequestFailed data) {
+    referSubscriber.on(EventReferRequestFailed(),
+        (EventReferRequestFailed data) {
       this._referSubscribers.remove(id);
     });
-    referSubscriber.on(EventAccepted(), (EventAccepted data) {
+    referSubscriber.on(EventReferAccepted(), (EventReferAccepted data) {
       this._referSubscribers.remove(id);
     });
-    referSubscriber.on(EventCallFailed(), (EventCallFailed data) {
+    referSubscriber.on(EventReferFailed(), (EventReferFailed data) {
       this._referSubscribers.remove(id);
     });
 
@@ -1959,8 +1961,7 @@ class RTCSession extends EventManager {
     // Reply before the transaction timer expires.
     request.reply(202);
 
-    var notifier =
-        new RTCSession_ReferNotifier.ReferNotifier(this, request.cseq);
+    var notifier = new ReferNotifier(this, request.cseq);
 
     var accept2 = (initCallback, options) {
       initCallback = (initCallback is Function) ? initCallback : null;
@@ -1977,16 +1978,16 @@ class RTCSession extends EventManager {
             event.response.status_code, event.response.reason_phrase);
       });
 
-      session.on(EventAccepted(), (EventAccepted event) {
+      session.on(EventCallAccepted(), (EventCallAccepted event) {
         notifier.notify(
             event.response.status_code, event.response.reason_phrase);
       });
 
       session.on(EventFailedUnderScore(), (EventFailedUnderScore data) {
-        if (data.message != null) {
-          notifier.notify(data.message.status_code, data.message.reason_phrase);
+        if (data.cause != null) {
+          notifier.notify(data.cause.status_code, data.cause.reason_phrase);
         } else {
-          notifier.notify(487, data.cause);
+          notifier.notify(487, data.cause.cause);
         }
       });
       // Consider the Replaces header present in the Refer-To URI.
@@ -2824,7 +2825,7 @@ class RTCSession extends EventManager {
     logger.debug('session accepted');
     this._start_time = new DateTime.now();
     logger.debug('emit "accepted"');
-    this.emit(EventAccepted(originator: originator, response: message));
+    this.emit(EventCallAccepted(originator: originator, response: message));
   }
 
   _confirmed(originator, ack) {
@@ -2855,7 +2856,6 @@ class RTCSession extends EventManager {
 
     this.emit(EventFailedUnderScore(
       originator: originator,
-      message: message,
       cause: errorCause,
     ));
 
@@ -2863,7 +2863,6 @@ class RTCSession extends EventManager {
     logger.debug('emit "failed"');
     this.emit(EventCallFailed(
         originator: originator,
-        message: message,
         request: request,
         cause: errorCause,
         response: response));
