+++ lib/src/RTCSession/ReferNotifier.dart
@@ -1,6 +1,7 @@
 import '../Constants.dart' as DartSIP_C;
 import '../Constants.dart';
 import '../event_manager/event_manager.dart';
+import '../event_manager/internal_events.dart';
 import '../logger.dart';
 
 class C {
