+++ lib/src/UA.dart
@@ -15,7 +15,9 @@ import 'Transport.dart';
 import 'URI.dart';
 import 'Utils.dart' as Utils;
 import 'WebSocketInterface.dart';
+
 import 'event_manager/event_manager.dart';
+import 'event_manager/internal_events.dart';
 import 'logger.dart';
 import 'sanityCheck.dart';
 import 'transactions/Transactions.dart';
