+++ lib/src/event_manager/event_manager.dart
@@ -1,8 +1,16 @@
-import '../../sip_ua.dart';
-import '../logger.dart';
 import 'events.dart';
+import 'call_events.dart';
+import 'refer_events.dart';
+import 'register_events.dart';
+import 'transport_events.dart';
 
 export 'events.dart';
+export 'call_events.dart';
+export 'refer_events.dart';
+export 'register_events.dart';
+export 'transport_events.dart';
+
+import '../logger.dart';
 
 /// This class serves as a Typed event bus.
 ///
