+++ lib/src/event_manager/events.dart
@@ -1,16 +1,3 @@
-import 'package:flutter_webrtc/webrtc.dart';
-
-import '../../sip_ua.dart';
-import '../NameAddrHeader.dart';
-import '../Message.dart';
-import '../RTCSession.dart';
-import '../RTCSession/DTMF.dart';
-import '../RTCSession/Info.dart';
-import '../SIPMessage.dart';
-import '../Transport.dart';
-import '../WebSocketInterface.dart';
-import '../transactions/transaction_base.dart';
-
 /// each EventType class can implement this method and the EventManager will call it before
 /// delivering an event, thus ensuring good quality events with a fail early approach.
 abstract class EventType {
@@ -31,26 +18,8 @@ abstract class EventType {
 /// These changes will make it much easier to reason about where Events go to and come from, as well as
 /// exactly what fields are available without the need to actually run the code.
 
-class EventStateChanged extends EventType {}
 
-class EventNewTransaction extends EventType {
-  // TransactionBase transaction;
-  EventNewTransaction({TransactionBase transaction});
-}
-
-class EventTransactionDestroyed extends EventType {
-//  TransactionBase transaction;
-  EventTransactionDestroyed({TransactionBase transaction});
-}
-
-class EventNewMessage extends EventType {
-  // String state;
-  //dynamic response;
-  String originator;
-  // MediaStream stream;
-  EventNewMessage({Message message, this.originator, OutgoingRequest request});
-}
-
+/// A general error cause class.
 class ErrorCause {
   int status_code;
   String cause;
@@ -60,322 +29,4 @@ class ErrorCause {
   String toString() {
     return 'Code: [$status_code], Cause: $cause, Reason: $reason_phrase';
   }
-}
-
-class EventRegistered extends EventType {
-  ErrorCause cause;
-  EventRegistered({this.cause});
-}
-
-class EventRegistrationFailed extends EventType {
-  ErrorCause cause;
-  EventRegistrationFailed({this.cause});
-}
-
-class EventUnregister extends EventType {
-  ErrorCause cause;
-  EventUnregister({this.cause});
-}
-
-class EventSipEvent extends EventType {
-  //OutgoingRequest request;
-  EventSipEvent({IncomingRequest request});
-}
-
-class EventSocketConnected extends EventType {
-  WebSocketInterface socket;
-  EventSocketConnected({this.socket});
-}
-
-class EventSocketConnecting extends EventType {
-  //OutgoingRequest request;
-  WebSocketInterface socket;
-
-  EventSocketConnecting({this.socket});
-}
-
-class EventSocketDisconnected extends EventType {
-  WebSocketInterface socket;
-  ErrorCause cause;
-  EventSocketDisconnected({WebSocketInterface socket, this.cause});
-}
-
-class EventStream extends EventType {
-  String originator;
-  MediaStream stream;
-  EventStream({this.originator, this.stream});
-}
-
-class EventOnAuthenticated extends EventType {
-  OutgoingRequest request;
-  EventOnAuthenticated({this.request});
-}
-
-class EventSdp extends EventType {
-//  String originator;
-//  String type;
-//  String sdp;
-  EventSdp({String originator, String type, String sdp});
-}
-
-class EventSending extends EventType {
-  // dynamic requset;
-  EventSending({OutgoingRequest request});
-}
-
-class EventSetRemoteDescriptionFailed extends EventType {
-  // dynamic exception;
-  EventSetRemoteDescriptionFailed({dynamic exception});
-}
-
-class EventSetLocalDescriptionFailed extends EventType {
-//  dynamic exception;
-  EventSetLocalDescriptionFailed({dynamic exception});
-}
-
-class EventFailedUnderScore extends EventType {
-//  String originator;
-  ErrorCause cause;
-  dynamic message;
-  EventFailedUnderScore({String originator, this.cause, this.message});
-}
-
-class EventGetUserMediaFailed extends EventType {
-//  dynamic exception;
-  EventGetUserMediaFailed({dynamic exception});
-}
-
-class EventNewDTMF extends EventType {
-//  String originator;
-  // OutgoingRequest request;
-  // DTMF dtmf;
-  EventNewDTMF({String originator, OutgoingRequest request, DTMF dtmf});
-}
-
-class EventNewInfo extends EventType {
-//  String originator;
-//  OutgoingRequest request;
-  // Info info;
-  EventNewInfo({String originator, OutgoingRequest request, Info info});
-}
-
-class EventPeerConnection extends EventType {
-//  RTCPeerConnection peerConnection;
-  EventPeerConnection(RTCPeerConnection peerConnection);
-}
-
-class EventReplaces extends EventType {
-//  OutgoingRequest request;
-//  bool Function(dynamic options) accept;
-//  bool Function(dynamic options) reject;
-  EventReplaces(
-      {String request,
-      bool Function(dynamic options) accept,
-      bool Function(dynamic options) reject});
-}
-
-class EventCallConfirmed extends EventType {
-//  String originator;
-//  String ack;
-  EventCallConfirmed({String originator, dynamic ack});
-}
-
-class EventUpdate extends EventType {
-//  OutgoingRequest request;
-//  bool Function(dynamic options) callback;
-//  bool Function(dynamic options) reject;
-  EventUpdate(
-      {IncomingRequest request,
-      bool Function(dynamic options) callback,
-      bool Function(dynamic options) reject});
-}
-
-class EventReinvite extends EventType {
-  // OutgoingRequest request;
-//  bool Function(dynamic options) callback;
-//  bool Function(dynamic options) reject;
-  EventReinvite(
-      {String request,
-      bool Function(dynamic options) callback,
-      bool Function(dynamic options) reject});
-}
-
-class EventIceCandidate extends EventType {
-//  RTCIceCandidate candidate;
-//  Future<Null> Function() ready;
-  EventIceCandidate(RTCIceCandidate candidate, Future<Null> Function() ready);
-}
-
-class EventCreateAnswerFialed extends EventType {
-//  dynamic exception;
-  EventCreateAnswerFialed({dynamic exception});
-}
-
-class EventCreateOfferFailed extends EventType {
-//  dynamic exception;
-  EventCreateOfferFailed({dynamic exception});
-}
-
-class EventCallRefer extends EventType {
-  String aor;
-  //bool Function({Function initCallback, dynamic options}) accept;
-  dynamic accept;
-  //bool Function(dynamic options) reject;
-  dynamic reject;
-  EventCallRefer({this.aor, this.accept, this.reject});
-}
-
-class EventCallConnecting extends EventType {
-  EventCallConnecting({dynamic request});
-}
-
-class EventCallEnded extends EventType {
-  String originator;
-  ErrorCause cause;
-  IncomingRequest request;
-//  dynamic message;
-  EventCallEnded({this.originator, this.cause, this.request});
-}
-
-class EventOnFialed extends EventType {}
-
-class EventTrying extends EventType {
-//  OutgoingRequest request;
-//  String status_line;
-  EventTrying({dynamic request, String status_line});
-}
-
-class EventCallProgress extends EventType {
-  // OutgoingRequest request;
-  // String status_line;
-  String originator;
-  dynamic response;
-  EventCallProgress(
-      {dynamic request, String status_line, this.originator, this.response});
-}
-
-class EventAccepted extends EventType {
-  // OutgoingRequest request;
-  // String status_line;
-  // String originator;
-  dynamic response;
-  EventAccepted(
-      {dynamic request, String status_line, String originator, this.response});
-}
-
-class EventCallAccepted extends EventType {}
-
-class EventCallFailed extends EventType {
-  // String state;
-  dynamic response;
-  String originator;
-  // MediaStream stream;
-  ErrorCause cause;
-  // dynamic message;
-  dynamic request;
-  // String status_line;
-  EventCallFailed(
-      {String state,
-      this.response,
-      this.originator,
-      MediaStream stream,
-      this.cause,
-      String message,
-      this.request,
-      String status_line});
-}
-
-class EventRequestSucceeded extends EventType {
-  //dynamic response;
-  EventRequestSucceeded({dynamic response});
-}
-
-class EventRequestFailed extends EventType {
-  //dynamic response;
-  //String cause;
-  EventRequestFailed({dynamic response, String cause});
-}
-
-class EventSucceeded extends EventType {
-  // String state;
-  // dynamic response;
-  // String originator;
-  // MediaStream stream;
-  // String cause;
-  EventSucceeded(
-      {String state,
-      IncomingMessage response,
-      String originator,
-      MediaStream stream,
-      String cause});
-}
-
-class EventGetusermediafailed extends EventType {
-  // dynamic exception;
-  EventGetusermediafailed({dynamic exception});
-}
-
-class EventCallHold extends EventType {
-  String originator;
-  EventCallHold({this.originator});
-}
-
-class EventCallUnhold extends EventType {
-  String originator;
-  EventCallUnhold({String originator});
-}
-
-class EventCallMuted extends EventType {
-  bool audio;
-  bool video;
-  EventCallMuted({this.audio, this.video});
-}
-
-class EventCallUnmuted extends EventType {
-  bool audio;
-  bool video;
-  EventCallUnmuted({this.audio, this.video});
-}
-
-class EventNewRTCSession extends EventType {
-  RTCSession session;
-  // String originator;
-  // OutgoingRequest request;
-  EventNewRTCSession({this.session, String originator, dynamic request});
-}
-
-class EventOnTransportError extends EventType {
-  EventOnTransportError() : super();
-}
-
-class EventOnRequestTimeout extends EventType {
-  IncomingMessage request;
-  EventOnRequestTimeout({this.request});
-}
-
-class EventOnReceiveResponse extends EventType {
-  IncomingResponse response;
-  EventOnReceiveResponse({this.response});
-  sanityCheck() {
-    assert(response != null);
-  }
-}
-
-class EventOnDialogError extends EventType {
-  IncomingMessage response;
-  EventOnDialogError({this.response});
-}
-
-class EventOnSuccessResponse extends EventType {
-  IncomingMessage response;
-  EventOnSuccessResponse({this.response});
-}
-
-class EventOnErrorResponse extends EventType {
-  IncomingMessage response;
-  EventOnErrorResponse({this.response});
-}
-
-class EventRegistrationExpiring extends EventType {
-  EventRegistrationExpiring();
 }
