+++ lib/src/event_manager/transport_events.dart
@@ -1,0 +1,18 @@
+import 'events.dart';
+import '../WebSocketInterface.dart';
+
+class EventSocketConnected extends EventType {
+  WebSocketInterface socket;
+  EventSocketConnected({this.socket});
+}
+
+class EventSocketConnecting extends EventType {
+  WebSocketInterface socket;
+  EventSocketConnecting({this.socket});
+}
+
+class EventSocketDisconnected extends EventType {
+  WebSocketInterface socket;
+  ErrorCause cause;
+  EventSocketDisconnected({WebSocketInterface socket, this.cause});
+}
