+++ lib/src/sip_ua_helper.dart
@@ -105,7 +105,13 @@ class SIPUAHelper extends EventManager {
 
   void refer(String target) {
     if (_session != null) {
-      _session.refer(target);
+      var refer = _session.refer(target);
+      refer.on(EventReferTrying(), (EventReferTrying data) {});
+      refer.on(EventReferProgress(), (EventReferProgress data) {});
+      refer.on(EventReferAccepted(), (EventReferAccepted data) {
+        _session.terminate();
+      });
+      refer.on(EventReferFailed(), (EventReferFailed data) {});
     }
   }
 
@@ -201,20 +207,17 @@ class SIPUAHelper extends EventManager {
       logger.debug('call connecting');
       _notifyCallStateListeners(CallState(CallStateEnum.CONNECTING));
     });
-
     eventHandlers.on(EventCallProgress(), (EventCallProgress event) {
       logger.debug('call is in progress');
       _notifyCallStateListeners(
           CallState(CallStateEnum.PROGRESS, originator: event.originator));
     });
-
     eventHandlers.on(EventCallFailed(), (EventCallFailed event) {
       logger.debug('call failed with cause: ' + (event.cause.toString()));
       _notifyCallStateListeners(CallState(CallStateEnum.FAILED,
           originator: event.originator, cause: event.cause));
       _session = null;
     });
-
     eventHandlers.on(EventCallEnded(), (EventCallEnded e) {
       logger.debug('call ended with cause: ' + (e.cause.toString()));
       _notifyCallStateListeners(CallState(CallStateEnum.ENDED,
@@ -257,10 +260,12 @@ class SIPUAHelper extends EventManager {
       });
     });
     eventHandlers.on(EventCallRefer(), (EventCallRefer refer) async {
-      logger.debug('call refer received');
+      logger.debug('Refer received, Transfer current call to => ${refer.aor}');
       _notifyCallStateListeners(CallState(CallStateEnum.REFER, refer: refer));
       //Always accept.
-      refer.accept(null, this._options(true));
+      refer.accept((session){
+        logger.debug('New session initialized.');
+      }, this._options(true));
     });
 
     var _defaultOptions = {
