+++ lib/src/transactions/invite_server.dart
@@ -4,6 +4,7 @@ import '../Timers.dart';
 import '../Transport.dart';
 import '../UA.dart';
 import '../event_manager/event_manager.dart';
+import '../event_manager/internal_events.dart';
 import '../logger.dart';
 import 'transaction_base.dart';
